Change the NewLabels signature.
Instead of accepting both an XYer and a Labeller, NewLabels
accepts an interface{ XYer; Labeller; }  This way we don't
have an explosion of one-off interfaces (LabelXYer, etc.)
and we don't need the XYer and Labeller to be passed
to NewLabels separately (requiring a run-time check to ensure
that they have the same Len()'s.

The only downside that I see of this approach is that the
Labeller interface can no longer have the Len() method since
there cannot be overlap in the methods of the composed
interfaces.  I don't believe this will be a problem down the
road, however, because you will always be labeling things
like XYer, Valuer, or XYZer, all of which have Len().

This sets a precedence for other Plotters like the (eventual)
XErrorBar, and YErrorBar plotters that will have XErrorer
and YErrorer interfaces that behave similarly to the Labeller
interface.  By the same argument made above, I don't think
that it will be a problem for these future interfaces to be
without a Len() method.