Implement Size() for the vg back-ends.
This method is used by NewDrawArea to get the size so that
it doesn't have to be specified twice.  Size() is not a required
method for vg.Canvas, but currently all of the back-ends are
finite-sized images, so it really cuts down on some annoying
double-specifying of image sizes.  A future, hypothetical,
unsized canvas (vggl, for OpenGL?) wouldn't implement this
method, and would use NewDrawAreaSize instead.