Fix label rounding for small tick value.
There are floating point errors in computing the tick value, which are
eliminated by rounding to a float32. If the value is close to zero, this
doesn't remove the rounding errors. If abs(value) < 1, add 1, convert to
float32, and then subtract one. This fixes cases where there are a lot
of digits for the axis label at 0.

Updates issue #18.