vg: drop DPI from Canvas
The DPI value is only ever used internally by types to determine the
number of dots for a given Length, the alternative is to give the Dots
method the DPI value rather than the Canvas.

The DPI method is retained on types that need to know their resolution,
although this could be made unexported.