Adding multiple Y axes.
The same thing could be made with the X axis, but it's far less common,
so I didn't do it.

The interface stays mostly the same. Instead of accessign plt.Y,
plt.Ys[0] (or any other index) must be used.

plt.AddYAxis() adds a new Y axis, and plt.AddWithAxis adds a plotter
with the given axis.

Tests have been updated. but only the svg and pdf golden files changed.
The differences are only in the order things are painted. The Y axis is
now painted before the x axis.

plotter/main has been updated with an example using 2 Y axes. More than
two axes are also possible of course.