all: add UnixTimeConverter and TimeFloatConverter
This CL introduces the TimeFloatConverter interface that describes how a
time.Time value is converted to a float64 (and back.)
A default TimeFloatConverter (used by TimeTicks) is also provided in the
shape of UnixTimeConverter.

Test updated to be resilient against timezones (tested with
'Europe/Paris', 'America/Los_Angeles' and 'Australia/Sydney')

Fixes #276.