Add PieChart plotter
This plotter draws pie charts, with slices of area proportional to the
Values given when initialized.

Color can be changed, labels can be added, offsets can be inserted
(similarly to what is done for bar charts).

Since this plotter is unrelated to axes, anybody using it will probably
want to HideAxes() too.