vg: restore FontDirs initialization
PR gonum/plot#335 moved the vg-syscall into a different file to support
gopherjs.
But it dropped the initialization of the FontDirs global variable.
Fix this and also move the global const 'importString' closer to its
usage, in vg/font_syscall.go.

Add a test to make sure we don't regress.