vg/...: invert image format imports
Image formats used by vg/draw need to be registered with
draw.RegisterFormat now. For this it was necessary to invert the imports
so that each implemenation also imports vg/draw.

A new package vgall was added which is used to import all image
formats at once. This can be disabled with the build tag "minimal".