vg/vgtex: implement Canvas.Fill without stroking path
We used to call pgfusepath with both 'stroke' and 'fill' when filling a
path.
This didn't show without transparency/opacity.
But with an opacity-enabled color, it does, as shown in go-hep/hep#618,
where the legend for a filled box is overlined with a not-closed
rectangle.

This CL now calls pgfusepath with only 'fill' as an argument.

Updates go-hep/hep#618.