text,vg/draw: extend draw.TextHandler interface
This CL extends the vg/draw.TextHandler interface to be able to handle
how lines are split, which may vary from backend to backend (e.g.: this
differs between text/plain and text/html).
TextHandler's Box method is also modified to follow x/image/font/sfnt
convention to return a positive number for the descent of a glyph
(a.k.a. "depth" in Box lingua.)

This CL also consolidates how multi-lines bounding boxes are computed,
leveraging the underlying draw.TextHandler of a given draw.TextStyle value.